Water_Probability/src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==2.4.1
autopep8==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lazy-object-proxy==1.4.3
MarkupSafe==2.1.5
mccabe==0.6.1
numpy==1.24.4
orjson==3.9.15
pandas==2.0.3
pycodestyle==2.6.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pylint==2.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rope==0.17.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
threadpoolctl==3.4.0
toml==0.10.1
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.12.1
